# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces": {
      "post": {
        "operationId": "WorkspacesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "SESSION_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "EMAIL_NOT_VERIFIED 또는 CSRF_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "WORKSPACE_LIMIT_REACHED 또는 WORKSPACE_SLUG_IN_USE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "첫 워크스페이스 생성",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspace": {
      "get": {
        "operationId": "WorkspacesController_getCurrent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "SESSION_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "EMAIL_NOT_VERIFIED 또는 MEMBERSHIP_INACTIVE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "현재 워크스페이스 조회",
        "tags": [
          "workspaces"
        ]
      }
    },
...
  },
  "info": {
    "title": "Rivet API",
    "description": "Rivet 비공개 MVP 베타 REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-rivet_session"
      }
    },
    "schemas": {
...
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "입력값을 확인해 주세요."
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "example": {
              "email": [
                "올바른 이메일 주소를 입력해 주세요."
              ]
            }
          },
          "requestId": {
            "type": "string",
            "example": "req_550e8400-e29b-41d4-a716-446655440000"
          },
          "currentVersion": {
            "type": "number",
            "minimum": 1
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message",
          "fieldErrors",
          "requestId"
        ]
      },
...
      "CreateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "example": "제품 개발팀",
            "minLength": 1
          },
          "slug": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "example": "product-team"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "WorkspaceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "제품 개발팀"
          },
          "slug": {
            "type": "string",
            "example": "product-team"
          },
          "version": {
            "type": "number",
            "example": 1,
            "minimum": 1
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "version"
        ]
      },
...
    }
  }
}